+++ frontend/src/routes/profile/chat/[id]/Chats.svelte
@@ -45,7 +45,7 @@
 		{#each $queryChats.data as { chat_id, chat_name, last_message_text, own }, i (chat_id)}
 			<li>
 				<a
-					href={route((p) => `/profile/chat/${p(chat_id)}`)}
+					href={route((p) => `/profile/chat/[id]/${p(chat_id)}`)}
 					class="group relative block overflow-hidden rounded-lg transition after:absolute after:-right-3 after:top-0 after:h-full after:w-3 after:bg-sky-900 after:transition-transform aria-[current]:translate-x-4 aria-[current]:bg-app-blue-100 aria-[current]:after:-translate-x-3"
 					aria-current={$page.params.id === chat_id || undefined}
 					bind:this={chatNodes[i]}
