+++ frontend/src/routes/profile/chat/[id]/[chatId]/+page.svelte
@@ -1,18 +1,20 @@
-<!-- <script lang="ts">
+<script lang="ts">
 	import { page } from '$app/stores';
 	import { createGetQuery } from '$lib/api';
-	import { createQuery, QueryObserver } from '@tanstack/svelte-query';
-	import { dateFormatter, type Chat, type Message, formatTime } from './interFace';
+	import { createQuery, useQueryClient } from '@tanstack/svelte-query';
+	import { dateFormatter, type Chat, type Message, formatTime } from '../interFace';
+	import { tick } from 'svelte';
 
-	// Функция для группировки массива по заданному ключу
-	function groupBy<T>(array: T[], keyFunc: (item: T) => string) {
-		return array.reduce((result: Record<string, T[]>, item) => {
+	const queryClient = useQueryClient();
+
+	const groupBy = <T,>(array: T[], keyFunc: (item: T) => string) =>
+		array.reduce((result: Record<string, T[]>, item) => {
 			const key = keyFunc(item);
-			result[key] = result[key] || [];
-			result[key].push(item);
+			result[key] = [...(result[key] || []), item];
 			return result;
 		}, {});
-	}
+
+	let newMessageText = '';
 
 	// Запрос на получение данных о чате
 	$: messagesGet = createGetQuery<Chat>(`/api/v1/chats/${$page.params.id || ''}`);
@@ -39,8 +41,15 @@
 		}
 	});
 
-	// Переменная для хранения текста нового сообщения
-	let newMessageText = '';
+	// Прокрутка вниз при получении новых сообщений
+  $: if ($queryMessage.isSuccess && $queryMessage.data) {
+    tick().then(() => {
+      const chatBottom = document.getElementById('chatBottom');
+      if (chatBottom) {
+        chatBottom.scrollIntoView({ behavior: 'smooth' });
+      }
+    });
+  }
 
 	// Функция для отправки нового сообщения
 	async function sendMessage() {
@@ -61,6 +70,9 @@
 			}
 
 			newMessageText = '';
+
+			// После успешной отправки сообщения вызываем invalidateQueries для обновления данных
+			queryClient.invalidateQueries(messagesGet.key as any);
 		} catch (error) {
 			console.error('Error sending message:', error);
 			// Обработка ошибки по необходимости
@@ -69,8 +81,8 @@
 </script>
 
 {#if $queryMessage.isSuccess && $queryMessage.data}
-	{@const messageGroups = groupBy($queryMessage.data.messages, (msg) =>
-		dateFormatter(msg.created_at)
+	{@const messageGroups = Object.entries(
+		groupBy($queryMessage.data.messages, (msg) => dateFormatter(msg.created_at))
 	)}
 
 	<div class="grid h-full grid-rows-5 border-2 border-l-0 border-sky-900">
@@ -80,13 +92,13 @@
 			</p>
 			<div class="overflow-y-hidden p-5">
 				<div class="h-full overflow-y-auto px-10 text-xl">
-					{#each Object.entries(messageGroups) as [date, messages] (date)}
+					{#each messageGroups as [date, messages]}
 						{#if messages.length > 0}
 							<p class="p-3 text-center">{date}</p>
 							<div class="flex flex-col gap-2">
 								{#each messages as msg (msg.created_at)}
 									{@const isYourMessage = msg.own}
-									<div class="max-w-lg {isYourMessage ? 'self-end' : ''}">
+									<div class="max-w-lg {isYourMessage && 'self-end'}">
 										<div
 											class="rounded-lg px-4 py-2 {isYourMessage
 												? 'bg-app-blue-50'
@@ -94,7 +106,7 @@
 										>
 											{msg.text}
 										</div>
-										<p class="text-neutral-700/50 {isYourMessage ? 'max-2xl text-right' : ''}">
+										<p class="text-neutral-700/50 {isYourMessage && 'max-2xl text-right'}">
 											{formatTime(msg.created_at)}
 										</p>
 									</div>
@@ -102,27 +114,29 @@
 							</div>
 						{/if}
 					{/each}
+					<div id="chatBottom"></div>
 				</div>
 			</div>
-			<div class="flex gap-2 px-16 py-6 font-title outline outline-2 outline-sky-900">
-				<div
-					class="grow rounded-3xl p-4 outline outline-1 outline-app-blue-600 focus-within:outline-2"
-				>
-					<textarea
-						bind:value={newMessageText}
-						placeholder="Type your message..."
-						class="mr-4 flex-grow rounded-lg border px-4 py-2 focus:outline-none"
-					/>
-				</div>
-				<button
-					on:click={sendMessage}
-					class="flex items-center justify-center rounded-3xl bg-app-blue-600 px-12 text-2xl text-white"
-				>
-					SEND
-				</button>
+		</div>
+		<div class="flex gap-2 px-16 py-6 font-title outline outline-2 outline-sky-900">
+			<div
+				class="grow rounded-3xl p-4 outline outline-1 outline-app-blue-600 focus-within:outline-2"
+			>
+				<textarea
+					bind:value={newMessageText}
+					placeholder="Type your message..."
+					class="w-full flex-grow rounded-lg focus:outline-none"
+					on:keydown={(event) => event.key === 'Enter' && sendMessage()}
+				/>
 			</div>
+			<button
+				on:click={sendMessage}
+				class="flex items-center justify-center rounded-3xl bg-app-blue-600 px-12 text-2xl text-white"
+			>
+				SEND
+			</button>
 		</div>
 	</div>
 {:else}
 	<div class="row-span-full m-auto">Loading...</div>
-{/if} -->
+{/if}
